port Minify_JS_ClosureCompiler test to phpunit
