+++ phpunit.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!-- http://phpunit.de/manual/current/en/appendixes.configuration.html -->
+<phpunit
+	backupGlobals="false"
+	backupStaticAttributes="false"
+	colors="true"
+	convertErrorsToExceptions="true"
+	convertNoticesToExceptions="true"
+	convertWarningsToExceptions="true"
+	processIsolation="false"
+	stopOnFailure="false"
+	syntaxCheck="false"
+	bootstrap="vendor/autoload.php">
+
+	<testsuites>
+		<testsuite name="Project Test Suite">
+			<directory>tests</directory>
+		</testsuite>
+	</testsuites>
+
+	<!--
+	<php>
+		<server name="KERNEL_DIR" value="/path/to/your/app/" />
+	</php>
+	-->
+
+	<!-- for code coverage -->
+	<filter>
+		<whitelist>
+			<directory>.</directory>
+			<exclude>
+				<directory>tests</directory>
+			</exclude>
+		</whitelist>
+	</filter>
+</phpunit>
